# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent API",
    "version": "1.0.0"
  },
...
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/taskslists": {
      "get": {
        "operationId": "tasksList",
        "parameters": [],
        "tags": [
          "tasks"
        ],
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "title"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createTaskList",
        "parameters": [],
        "tags": [
          "tasks"
        ],
        "description": "Create a new task list for the authenticated user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "title"
                ],
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response for 201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "userId",
                    "title",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          }
        }
      }
    },
...
    "/tasks/tree": {
      "get": {
        "operationId": "tasksTree",
        "parameters": [],
        "tags": [
          "tasks"
        ],
        "description": "Return all task lists with their active tasks and subtasks for the authenticated user.",
        "responses": {
          "200": {
            "description": "Response for 200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "allOf": [
                      {
                        "type": "object",
                        "properties": {
                          "tasks": {
                            "type": "array",
                            "items": {
                              "allOf": [
                                {
                                  "type": "object",
                                  "properties": {
                                    "subtasks": {
                                      "type": "array",
                                      "items": {
                                        "type": "object",
                                        "properties": {
                                          "status": {
                                            "type": "string"
                                          },
                                          "id": {
                                            "type": "string"
                                          },
                                          "title": {
                                            "type": "string"
                                          },
                                          "updatedAt": {
                                            "type": "string"
                                          },
                                          "notes": {
                                            "anyOf": [
                                              {
                                                "type": "null"
                                              },
                                              {
                                                "type": "string"
                                              }
                                            ]
                                          },
                                          "due": {
                                            "anyOf": [
                                              {
                                                "type": "null"
                                              },
                                              {
                                                "type": "string"
                                              }
                                            ]
                                          },
                                          "deletedAt": {
                                            "anyOf": [
                                              {
                                                "type": "null"
                                              },
                                              {
                                                "type": "string"
                                              }
                                            ]
                                          },
                                          "taskId": {
                                            "type": "string"
                                          },
                                          "mediaId": {
                                            "anyOf": [
                                              {
                                                "type": "null"
                                              },
                                              {
                                                "type": "string"
                                              }
                                            ]
                                          }
                                        },
                                        "required": [
                                          "status",
                                          "id",
                                          "title",
                                          "updatedAt",
                                          "notes",
                                          "due",
                                          "deletedAt",
                                          "taskId",
                                          "mediaId"
                                        ],
                                        "additionalProperties": false
                                      }
                                    }
                                  },
                                  "required": [
                                    "subtasks"
                                  ],
                                  "additionalProperties": false
                                },
                                {
                                  "type": "object",
                                  "properties": {
                                    "status": {
                                      "type": "string"
                                    },
                                    "id": {
                                      "type": "string"
                                    },
                                    "title": {
                                      "type": "string"
                                    },
                                    "updatedAt": {
                                      "type": "string"
                                    },
                                    "notes": {
                                      "anyOf": [
                                        {
                                          "type": "null"
                                        },
                                        {
                                          "type": "string"
                                        }
                                      ]
                                    },
                                    "due": {
                                      "anyOf": [
                                        {
                                          "type": "null"
                                        },
                                        {
                                          "type": "string"
                                        }
                                      ]
                                    },
                                    "taskListId": {
                                      "type": "string"
                                    },
                                    "deletedAt": {
                                      "anyOf": [
                                        {
                                          "type": "null"
                                        },
                                        {
                                          "type": "string"
                                        }
                                      ]
                                    }
                                  },
                                  "required": [
                                    "status",
                                    "id",
                                    "title",
                                    "updatedAt",
                                    "notes",
                                    "due",
                                    "taskListId",
                                    "deletedAt"
                                  ],
                                  "additionalProperties": false
                                }
                              ]
                            }
                          }
                        },
                        "required": [
                          "tasks"
                        ],
                        "additionalProperties": false
                      },
                      {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "userId": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "userId",
                          "title",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{taskId}/subtasks": {
      "post": {
        "operationId": "createSubtask",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "tasks"
        ],
        "description": "Create a new subtask under a task owned by the authenticated user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "title"
                ],
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "notes": {
                    "type": "string"
                  },
                  "due": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Response for 201",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "notes": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "string"
                        }
                      ]
                    },
                    "due": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "string"
                        }
                      ]
                    },
                    "deletedAt": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "string"
                        }
                      ]
                    },
                    "taskId": {
                      "type": "string"
                    },
                    "mediaId": {
                      "anyOf": [
                        {
                          "type": "null"
                        },
                        {
                          "type": "string"
                        }
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "id",
                    "title",
                    "updatedAt",
                    "notes",
                    "due",
                    "deletedAt",
                    "taskId",
                    "mediaId"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ValidationError"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedErr"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UnauthorizedErr": {
        "type": "object",
        "required": [
          "type",
          "title"
        ],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "Unauthorized"
            ]
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}